bots/arxiv-bot/requirements.txt
feedparser
holidays
slackclient
